# server/docs/swag/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is api document for Guard",
        "title": "Swagger Guard API",
        "termsOfService": "https://github.com/sysarmor/guard",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/guard/ca": {
            "get": {
                "description": "Get CA certificate",
                "tags": [
                    "Guard"
                ],
                "summary": "GetCA",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "unix timestamp, seconds",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CA certificate",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/guard/krl": {
            "get": {
                "description": "Get revoked key list (KRL)",
                "tags": [
                    "Guard"
                ],
                "summary": "GetKRL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "unix timestamp, seconds",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "KRL",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/guard/principals": {
            "get": {
                "description": "Get principals",
                "tags": [
                    "Guard"
                ],
                "summary": "GetPrincipals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "unix timestamp, seconds",
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "signature",
                        "name": "X-Signature",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "principals",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Principals"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/guard/space": {
            "get": {
                "description": "List spaces",
                "tags": [
                    "space"
                ],
                "summary": "ListSpace",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.ListSpaceVO"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create space",
                "tags": [
                    "space"
                ],
                "summary": "CreateSpace",
                "parameters": [
                    {
                        "description": "Create space request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateSpaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/guard/space/{spaceID}/node": {
            "get": {
                "description": "List nodes",
                "tags": [
                    "node"
                ],
                "summary": "ListNode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ListNodeResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create node",
                "tags": [
                    "node"
                ],
                "summary": "CreateNode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create node request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateNodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.CreateNodeResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/guard/space/{spaceID}/node/{nodeID}": {
            "delete": {
                "description": "Delete node",
                "tags": [
                    "node"
                ],
                "summary": "DeleteNode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Node ID",
                        "name": "nodeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/guard/space/{spaceID}/role": {
            "get": {
                "description": "List roles",
                "tags": [
                    "role"
                ],
                "summary": "ListRole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.ListRoleVO"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/guard/space/{spaceID}/role/{roleID}": {
            "delete": {
                "description": "Delete role",
                "tags": [
                    "role"
                ],
                "summary": "DeleteRole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/guard/space/{spaceID}/role/{roleID}/node": {
            "get": {
                "description": "List role nodes",
                "tags": [
                    "role"
                ],
                "summary": "ListRoleNode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.RoleNodeListVO"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add node to role",
                "tags": [
                    "role"
                ],
                "summary": "AddNodeToRole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "node list",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.RoleNodeRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/guard/space/{spaceID}/role/{roleID}/node/batch/delete": {
            "post": {
                "description": "Batch remove node from role",
                "tags": [
                    "role"
                ],
                "summary": "BatchRemoveNodeFromRole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Node IDs",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/guard/space/{spaceID}/role/{roleID}/user/batch/delete": {
            "post": {
                "description": "Batch remove user from role",
                "tags": [
                    "role"
                ],
                "summary": "BatchRemoveUserFromRole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "spaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User IDs",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/guard/user/{userID}": {
            "get": {
                "description": "Get user",
                "tags": [
                    "user"
                ],
                "summary": "GetUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.UserVO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.Principals": {
            "type": "object",
            "properties": {
                "principals": {
                    "description": "Principals are the principals of the role\nthe principals are the email of the users",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role": {
                    "description": "Role is the role of the principals",
                    "type": "string"
                }
            }
        },
...
        "service.CreateSpaceRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "service.ListNodeResponse": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.ListNodeVO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "service.ListSpaceVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}